Profile refactoring: renaming files and folders
