+++ src/components/Content/Profile/Profile.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import css from './Profile.module.css';
-import ProfileInfo from "./ProfileInfo";
-import MyPostsContainer from './MyPostsContainer';
+import ProfileInfo from "./ProfileInfo/ProfileInfo";
+import MyPostsContainer from './MyPost/MyPostsContainer';
 import background from '../../../assets/images/background.png'
 import { ProfilePropsType } from "./ProfileContainer";
