+++ src/components/Content/Users/User.tsx
@@ -9,10 +9,7 @@ type PropsType = {
     followingInProgress: Array<number>
     follow: (id: number) => void
     unfollow: (id: number) => void
-
 }
-
-
 
 const User: React.FC<PropsType> = ({ user, followingInProgress, unfollow, follow }) => {
 
@@ -47,8 +44,6 @@ const User: React.FC<PropsType> = ({ user, followingInProgress, unfollow, follow
                 }
             </div>
         </div>
-
-
     )
 
 }
